feat: become mode
