+++ packages/vtable-plugins/src/master-detail-plugin/index.ts
@@ -1,0 +1,5 @@
+import { MasterDetailPlugin } from './core';
+import type { MasterDetailPluginOptions, DetailGridOptions } from './types';
+
+export { MasterDetailPlugin };
+export type { MasterDetailPluginOptions, DetailGridOptions };
