+++ packages/vtable-plugins/src/master-detail-plugin/rendering/index.ts
@@ -1,0 +1,151 @@
+import * as VTable from '@visactor/vtable';
+import { getRowCells } from '../utils';
+
+/**
+ * 渲染管理器类
+ */
+export class RenderingManager {
+  private table: VTable.ListTable;
+  private expandedRows: number[];
+
+  constructor(table: VTable.ListTable, expandedRows: number[]) {
+    this.table = table;
+    this.expandedRows = expandedRows;
+  }
+
+  /**
+   * 为指定行绘制下划线
+   */
+  drawUnderlineForRow(rowIndex: number, originalHeight: number): void {
+    const sceneGraph = (this.table as any).scenegraph;
+    if (!sceneGraph) {
+      return;
+    }
+    // 获取指定行的所有cell
+    const rowCells = getRowCells(this.table, rowIndex);
+    if (rowCells.length === 0) {
+      return;
+    }
+    rowCells.forEach((cellGroup: any, index: number) => {
+      if (cellGroup && cellGroup.attribute) {
+        // 为这个cell添加下划线渲染逻辑
+        this.addUnderlineToCell(cellGroup, originalHeight);
+      }
+    });
+    // 触发重新渲染
+    this.table.scenegraph.updateNextFrame();
+  }
+
+  /**
+   * 为cell添加下划线
+   */
+  private addUnderlineToCell(cellGroup: any, originalHeight: number): void {
+    // 在CellGroup重绘时正确绘制下划线
+    const currentAttr = cellGroup.attribute;
+    const currentStrokeArrayWidth =
+      currentAttr.strokeArrayWidth ||
+      (currentAttr.lineWidth
+        ? [currentAttr.lineWidth, currentAttr.lineWidth, currentAttr.lineWidth, currentAttr.lineWidth]
+        : [1, 1, 1, 1]);
+    const currentStrokeArrayColor =
+      currentAttr.strokeArrayColor ||
+      (currentAttr.stroke
+        ? [currentAttr.stroke, currentAttr.stroke, currentAttr.stroke, currentAttr.stroke]
+        : ['transparent', 'transparent', 'transparent', 'transparent']);
+    const isAlreadyEnhanced = cellGroup._hasUnderline;
+    if (!isAlreadyEnhanced) {
+      // 第一次添加下划线，存储原始样式
+      cellGroup._originalStrokeArrayWidth = [...currentStrokeArrayWidth];
+      cellGroup._originalStrokeArrayColor = [...currentStrokeArrayColor];
+      cellGroup._hasUnderline = true;
+    }
+    // 使用原始样式计算增强的下划线
+    const originalStrokeArrayWidth = cellGroup._originalStrokeArrayWidth || currentStrokeArrayWidth;
+    const originalStrokeArrayColor = cellGroup._originalStrokeArrayColor || currentStrokeArrayColor;
+    const enhancedStrokeArrayWidth = [...originalStrokeArrayWidth];
+    const enhancedStrokeArrayColor = [...originalStrokeArrayColor];
+    const originalBottomWidth = originalStrokeArrayWidth[2] || 1;
+    const dpr = (this.table as any).internalProps?.pixelRatio || window.devicePixelRatio || 1;
+    // 要还原本来的下划线的效果，那么我们应该要 * 1.5因为我记得原本的线是叠层的
+    const enhancedWidth = originalBottomWidth * 1.5 * dpr;
+    enhancedStrokeArrayWidth[2] = Math.max(enhancedWidth, 1.5 * dpr);
+    if (originalStrokeArrayColor[2] === 'transparent' || !originalStrokeArrayColor[2]) {
+      const theme = (this.table as any).theme;
+      enhancedStrokeArrayColor[2] = theme?.bodyStyle?.borderColor || '#e1e4e8';
+    } else {
+      enhancedStrokeArrayColor[2] = originalStrokeArrayColor[2];
+    }
+    cellGroup.setAttributes({
+      strokeArrayWidth: enhancedStrokeArrayWidth,
+      strokeArrayColor: enhancedStrokeArrayColor,
+      stroke: true
+    });
+  }
+
+  /**
+   * 删除展开行的下划线
+   */
+  removeUnderlineFromRow(rowIndex: number): void {
+    // 获取指定行的所有cell
+    const rowCells = getRowCells(this.table, rowIndex);
+    rowCells.forEach((cellGroup: any, index: number) => {
+      if (cellGroup && cellGroup._hasUnderline) {
+        this.removeUnderlineFromCell(cellGroup);
+      }
+    });
+    // 触发重新渲染
+    this.table.scenegraph.updateNextFrame();
+  }
+
+  /**
+   * 从cell中删除下划线，恢复原始strokeArray样式
+   */
+  private removeUnderlineFromCell(cellGroup: any): void {
+    if (cellGroup._hasUnderline) {
+      // 恢复原始的strokeArray样式
+      if (cellGroup._originalStrokeArrayWidth && cellGroup._originalStrokeArrayColor) {
+        cellGroup.setAttributes({
+          strokeArrayWidth: cellGroup._originalStrokeArrayWidth,
+          strokeArrayColor: cellGroup._originalStrokeArrayColor
+        });
+        // 清理存储的原始样式
+        delete cellGroup._originalStrokeArrayWidth;
+        delete cellGroup._originalStrokeArrayColor;
+      }
+      // 清除下划线标记
+      cellGroup._hasUnderline = false;
+    }
+  }
+
+  /**
+   * 重绘所有展开行的下划线
+   */
+  redrawAllUnderlines(getOriginalRowHeight: (bodyRowIndex: number) => number): void {
+    this.expandedRows.forEach(rowIndex => {
+      const bodyRowIndex = rowIndex - this.table.columnHeaderLevelCount;
+      const originalHeight = getOriginalRowHeight(bodyRowIndex);
+      if (originalHeight > 0) {
+        // 先删除旧的下划线
+        this.removeUnderlineFromRow(rowIndex);
+        // 重新绘制下划线
+        this.drawUnderlineForRow(rowIndex, originalHeight);
+      }
+    });
+  }
+
+  /**
+   * 刷新指定行的图标状态
+   */
+  refreshRowIcon(rowIndex: number): void {
+    try {
+      const firstColumnIndex = 0;
+      const sceneGraph = (this.table as any).scenegraph;
+      if (sceneGraph && sceneGraph.updateCellContent) {
+        sceneGraph.updateCellContent(firstColumnIndex, rowIndex);
+        sceneGraph.updateNextFrame();
+      }
+    } catch (error) {
+      console.warn('刷新图标状态失败:', error);
+    }
+  }
+}
